requirements.txt
altair==5.3.0
APScheduler==3.6.3
blinker==1.7.0
cachetools==4.2.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
entrypoints==0.4
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
importlib-metadata==7.1.0
jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
numpy<2,>=1.22.4
packaging==24.0
pandas==2.3.0
Pillow==10.3.0
psutil==5.9.8
pyarrow==15.0.2
Pygments==2.18.0
pympler==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==13.15
pytz==2025.2
requests==2.32.4
rich==13.7.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.17.0
smmap==5.0.1
streamlit==1.35.0
ta==0.11.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.1
tqdm==4.66.4
tzdata==2025.2
tzlocal==5.3.1
typing-extensions==4.12.1
urllib3==2.2.1
validators==0.28.1
watchdog==4.0.1
zipp==3.18.2
loguru==0.7.2
seaborn>=0.12.2